pyproject.toml
[tool.poetry]
name = "goskey-signature-collector"
version = "0.0.0"
description = "Сервис для сбора подписей под массовыми обращениями в госорганы"
authors = ["volodkindv <volodkindv@yandex.ru>"]
readme = "README.md"
packages = [{include = "src", from = "."}]


[tool.poetry.dependencies]
python = "3.10.*"

pydantic = "^2.5.2"
cryptography = "^41.0.7"
odswriter = "^0.4.0"
aiogram = "^3.3.0"
pydantic-settings = "^2.1.0"
anyio = "^4.2.0"

[tool.poetry.group.lint.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
flake8-pyproject = "^1.2.3"
flake8-logging-format = "^0.9.0"
flake8-dunder-all = "^0.3.0"
flake8-bandit = "^4.1.1"
wemake-python-styleguide = "^0.18.0"
bandit = {extras = ["toml"], version = "^1.7.5"}
mypy = "^1.5.1"
types-requests = "^2.31.0.5"

[tool.poetry.group.test.dependencies]
requests-mock = "^1.11.0"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-dependency = "^0.5.1"
coverage = "^7.3.1"
coverage-badge = "^1.1.0"
factory_boy = "^3.2.1"
pytest-xdist = "^3.4.0"
polyfactory = "^2.13.0"

[tool.poetry.group.dev.dependencies]
poetry-plugin-export = "^1.6.0"

[tool.flake8]
exclude = []
max-line-length = 119
extend-ignore = [
    "D",  # flake8-docstrings
    "RST",  # docstrings
    "E501",  # Line too long
    "WPS115",  # Require `snake_case` for naming class attributes
    "WPS226",  # Forbid the overuse of string literals
    "WPS305",  # Forbid f strings
    "WPS306",  # Forbid writing classes without base classes
    "WPS323",  # Forbid % formatting on strings
    "WPS332",  # Forbid walrus operator
    "WPS404",  # Forbid complex defaults
    "WPS410",  # Forbid some module-level variables
    "WPS420",  # Forbid some python keywords
    "WPS431",  # Forbid nested classes
    "WPS432",  # Forbid magic numbers
    "WPS604",  # Forbid incorrect nodes inside class definitions
    "WPS300",  # Found local folder import
    "WPS237", # Found a too complex `f` string
    "WPS412", # Found `__init__.py` module with logic
]

per-file-ignores = [
#    "src/api_schemas.py:N815,WPS110,WPS124,E800",  # контракты внешнего сервиса
]

enable-extensions = ["G"]
inline-quotes = "double"
extend-immutable-calls = [
    "Path",
    "Query",
    "Header",
    "Cookie",
    "Body",
    "Form",
    "File",
]
max-imports = 17
max-module-members = 15
max-arguments = 9
max-local-variables = 9
max-returns = 9
max-line-complexity = 21

[tool.black]
line-length = 119

[tool.isort]
line_length = 119
profile = "black"

[tool.mypy]
mypy_path = "src"
exclude = [
    # "src/example.py",
]
explicit_package_bases = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_site_packages = true
plugins = [
    "pydantic.mypy",
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
pythonpath = ["tests"]
addopts = [
#    "--import-mode=importlib",
     "-p no:warnings",
#     "--no-cov",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
